+++ apps/opik-backend/src/main/java/com/comet/opik/infrastructure/aws/AwsModule.java
@@ -4,12 +4,19 @@ import com.comet.opik.infrastructure.OpikConfiguration;
 import com.comet.opik.infrastructure.S3Config;
 import com.google.inject.Provides;
 import jakarta.inject.Singleton;
+import lombok.NonNull;
 import ru.vyarus.dropwizard.guice.module.support.DropwizardAwareModule;
 import ru.vyarus.dropwizard.guice.module.yaml.bind.Config;
 import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
 import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
 import software.amazon.awssdk.auth.credentials.ContainerCredentialsProvider;
 import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.s3.S3Configuration;
+import software.amazon.awssdk.services.s3.presigner.S3Presigner;
+
+import java.net.URI;
 
 public class AwsModule extends DropwizardAwareModule<OpikConfiguration> {
 
@@ -21,5 +28,43 @@ public class AwsModule extends DropwizardAwareModule<OpikConfiguration> {
         }
 
         return StaticCredentialsProvider.create(AwsBasicCredentials.create(config.getS3Key(), config.getS3Secret()));
+    }
+
+    @Provides
+    @Singleton
+    public S3Client s3Client(@Config("s3Config") S3Config config, @NonNull AwsCredentialsProvider credentialsProvider) {
+        Region region = Region.of(config.getS3Region());
+
+        var builder = S3Client.builder()
+                .region(region)
+                .credentialsProvider(credentialsProvider);
+
+        if (config.isMinIO()) {
+            builder.forcePathStyle(true)
+                    .endpointOverride(URI.create(config.getS3Url()));
+        }
+
+        return builder.build();
+    }
+
+    @Provides
+    @Singleton
+    public S3Presigner preSigner(@Config("s3Config") S3Config config,
+            @NonNull AwsCredentialsProvider credentialsProvider) {
+        Region region = Region.of(config.getS3Region());
+        S3Configuration s3Configuration = S3Configuration.builder()
+                .pathStyleAccessEnabled(true)
+                .build();
+
+        var builder = S3Presigner.builder()
+                .credentialsProvider(credentialsProvider)
+                .region(region)
+                .serviceConfiguration(s3Configuration);
+
+        if (config.isMinIO()) {
+            builder.endpointOverride(URI.create(config.getS3Url()));
+        }
+
+        return builder.build();
     }
 }
